+++ code/src/components/Loading.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Loading = () => {
+  return (
+    <p>Loading</p>
+  )
+}
+
+export default Loading;
